pyproject.toml
[tool.ruff]
line-length = 128

[project]
name = "threebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "eiogram>=1.7.2",
    "fastapi==0.115.13",
    "python-decouple>=3.8",
    "sqlalchemy[asyncio]>=2.0.42",
    "uvicorn>=0.35.0",
]
